Issue #2947 use the existing utility class for manipulating focus
