+++ app/databrowser/src/main/java/org/csstudio/trends/databrowser3/ui/properties/TracesTab.java
@@ -37,7 +37,7 @@ import org.phoebus.framework.selection.SelectionService;
 import org.phoebus.ui.application.ContextMenuHelper;
 import org.phoebus.ui.dialog.AlertWithToggle;
 import org.phoebus.ui.dialog.DialogHelper;
-import org.phoebus.ui.focus.FocusUtility;
+import org.phoebus.ui.javafx.FocusUtil;
 import org.phoebus.ui.undo.UndoableActionManager;
 import org.phoebus.util.time.SecondsParser;
 
@@ -742,7 +742,7 @@ public class TracesTab extends Tab
             if (pvs.size() > 0)
             {
                 SelectionService.getInstance().setSelection(this, pvs);
-                ContextMenuHelper.addSupportedEntries(FocusUtility.setFocusOn(trace_table), menu);
+                ContextMenuHelper.addSupportedEntries(FocusUtil.setFocusOn(trace_table), menu);
             }
 
             menu.show(trace_table.getScene().getWindow(), event.getScreenX(), event.getScreenY());
