+++ app/pvtable/src/main/java/org/phoebus/applications/pvtable/ui/PVTable.java
@@ -35,7 +35,7 @@ import org.phoebus.ui.autocomplete.PVAutocompleteMenu;
 import org.phoebus.ui.dialog.DialogHelper;
 import org.phoebus.ui.dialog.NumericInputDialog;
 import org.phoebus.ui.dnd.DataFormats;
-import org.phoebus.ui.focus.FocusUtility;
+import org.phoebus.ui.javafx.FocusUtil;
 import org.phoebus.ui.javafx.PrintAction;
 import org.phoebus.ui.javafx.Screenshot;
 import org.phoebus.ui.javafx.ToolbarHelper;
@@ -692,7 +692,7 @@ public class PVTable extends VBox
             }
 
             // Add PV entries
-            if (ContextMenuHelper.addSupportedEntries(FocusUtility.setFocusOn(table), menu))
+            if (ContextMenuHelper.addSupportedEntries(FocusUtil.setFocusOn(table), menu))
                 menu.getItems().add(new SeparatorMenuItem());
 
             menu.getItems().add(new PrintAction(this));
