+++ app/pvtree/src/main/java/org/phoebus/applications/pvtree/ui/FXTree.java
@@ -28,7 +28,7 @@ import org.phoebus.framework.selection.SelectionService;
 import org.phoebus.ui.application.ContextMenuHelper;
 import org.phoebus.ui.application.ContextMenuService;
 import org.phoebus.ui.application.SaveSnapshotAction;
-import org.phoebus.ui.focus.FocusUtility;
+import org.phoebus.ui.javafx.FocusUtil;
 import org.phoebus.ui.javafx.PrintAction;
 import org.phoebus.ui.javafx.Screenshot;
 import org.phoebus.ui.javafx.TreeHelper;
@@ -139,7 +139,7 @@ public class FXTree
         {
             menu.getItems().clear();
 
-            if (ContextMenuHelper.addSupportedEntries(FocusUtility.setFocusOn(tree_view), menu))
+            if (ContextMenuHelper.addSupportedEntries(FocusUtil.setFocusOn(tree_view), menu))
                 menu.getItems().add(new SeparatorMenuItem());
             menu.getItems().add(new PrintAction(tree_view));
             menu.getItems().add(new SaveSnapshotAction(tree_view));
